+++ sonaranalyzer-dotnet/src/SonarAnalyzer.CSharp/Rules/MethodParameterUnused.cs
@@ -27,8 +27,8 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Diagnostics;
 using SonarAnalyzer.Common;
 using SonarAnalyzer.Helpers;
-using SonarAnalyzer.SymbolicExecution.CFG;
-using SonarAnalyzer.SymbolicExecution.LVA;
+using SonarAnalyzer.SymbolicExecution.ControlFlowGraph;
+using SonarAnalyzer.SymbolicExecution.LiveVariableAnalysis;
 
 namespace SonarAnalyzer.Rules.CSharp
 {
