+++ sonaranalyzer-dotnet/src/SonarAnalyzer.CSharp/SymbolicExecution/ControlFlowGraph/Blocks/ForeachCollectionProducerBlock.cs
@@ -21,7 +21,7 @@
 using System;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-namespace SonarAnalyzer.SymbolicExecution.CFG
+namespace SonarAnalyzer.SymbolicExecution.ControlFlowGraph
 {
     public sealed class ForeachCollectionProducerBlock : SimpleBlock
     {
