+++ sonaranalyzer-dotnet/src/SonarAnalyzer.Common/SymbolicExecution/ControlFlowGraph/SyntaxNodeBased/AbstractControlFlowGraphBuilder.cs
@@ -24,7 +24,7 @@ using System.Collections.Immutable;
 using System.Linq;
 using Microsoft.CodeAnalysis;
 
-namespace SonarAnalyzer.SymbolicExecution.CFG
+namespace SonarAnalyzer.SymbolicExecution.ControlFlowGraph
 {
     internal abstract class AbstractControlFlowGraphBuilder
     {
