+++ sonaranalyzer-dotnet/src/SonarAnalyzer.Common/SymbolicExecution/ControlFlowGraph/SyntaxNodeBased/Blocks/ExitBlock.cs
@@ -18,7 +18,7 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-namespace SonarAnalyzer.SymbolicExecution.CFG
+namespace SonarAnalyzer.SymbolicExecution.ControlFlowGraph
 {
     public sealed class ExitBlock : Block
     {
