+++ sonaranalyzer-dotnet/src/SonarAnalyzer.Common/SymbolicExecution/ControlFlowGraph/SyntaxNodeBased/Blocks/JumpBlock.cs
@@ -22,7 +22,7 @@ using System;
 using System.Collections.Generic;
 using Microsoft.CodeAnalysis;
 
-namespace SonarAnalyzer.SymbolicExecution.CFG
+namespace SonarAnalyzer.SymbolicExecution.ControlFlowGraph
 {
     public sealed class JumpBlock : SimpleBlock
     {
