+++ sonaranalyzer-dotnet/src/SonarAnalyzer.Common/SymbolicExecution/LiveVariableAnalysis/AbstractLiveVariableAnalysis.cs
@@ -23,9 +23,9 @@ using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
 using Microsoft.CodeAnalysis;
-using SonarAnalyzer.SymbolicExecution.CFG;
+using SonarAnalyzer.SymbolicExecution.ControlFlowGraph;
 
-namespace SonarAnalyzer.SymbolicExecution
+namespace SonarAnalyzer.SymbolicExecution.LiveVariableAnalysis
 {
     public abstract class AbstractLiveVariableAnalysis
     {
